+++ picasso_workflow/tests/test_confluence.py
@@ -402,6 +402,22 @@ class Test_B_ConfluenceReporterModules(unittest.TestCase):
         )
         self.cr.ci.delete_page(pgid)
 
+    def register_channels(self):
+        parameters = {}
+        results = {
+            "start time": "now",
+            "duration": 1.3,
+            "registration_model": "affine",
+            "registration_rms": [0.1, 0.2],
+        }
+        self.cr.register_channels(0, parameters, results)
+
+        # clean up
+        pgid, pgtitle = self.cr.ci.get_page_properties(
+            self.cr.report_page_name
+        )
+        self.cr.ci.delete_page(pgid)
+
     def convert_zeiss_movie(self):
         parameters = {"filepath": "myfile.czi"}
         results = {